[Ref] Symmod: Silence GCC -Wconversion warnings for float conversions.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19263 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_symmod.cpp
@@ -1779,28 +1779,28 @@ bool CSoundFile::ReadSymMOD(FileReader &file, ModLoadingFlags loadFlags)
 								if(patternSpeed > 1 && chnState.curVolSlideAmt >= (patternSpeed - 1))
 								{
 									uint8 slideAmt = std::min<uint8>(15, mpt::saturate_round<uint8>(chnState.curVolSlideAmt / (patternSpeed - 1)));
-									chnState.curVolSlideAmt -= slideAmt * (patternSpeed - 1);
+									chnState.curVolSlideAmt -= static_cast<float>(slideAmt * (patternSpeed - 1));
 									// normal slide up
 									m.command = CMD_CHANNELVOLSLIDE;
 									m.param = slideAmt << 4;
 								} else if(chnState.curVolSlideAmt >= 1.0f)
 								{
 									uint8 slideAmt = std::min<uint8>(15, mpt::saturate_round<uint8>(chnState.curVolSlideAmt));
-									chnState.curVolSlideAmt -= slideAmt;
+									chnState.curVolSlideAmt -= static_cast<float>(slideAmt);
 									// fine slide up
 									m.command = CMD_CHANNELVOLSLIDE;
 									m.param = (slideAmt << 4) | 0x0F;
 								} else if(patternSpeed > 1 && chnState.curVolSlideAmt <= -(patternSpeed - 1))
 								{
 									uint8 slideAmt = std::min<uint8>(15, mpt::saturate_round<uint8>(-chnState.curVolSlideAmt / (patternSpeed - 1)));
-									chnState.curVolSlideAmt += slideAmt * (patternSpeed - 1);
+									chnState.curVolSlideAmt += static_cast<float>(slideAmt * (patternSpeed - 1));
 									// normal slide down
 									m.command = CMD_CHANNELVOLSLIDE;
 									m.param = slideAmt;
 								} else if(chnState.curVolSlideAmt <= -1.0f)
 								{
 									uint8 slideAmt = std::min<uint8>(14, mpt::saturate_round<uint8>(-chnState.curVolSlideAmt));
-									chnState.curVolSlideAmt += slideAmt;
+									chnState.curVolSlideAmt += static_cast<float>(slideAmt);
 									// fine slide down
 									m.command = CMD_CHANNELVOLSLIDE;
 									m.param = slideAmt | 0xF0;
@@ -1816,28 +1816,28 @@ bool CSoundFile::ReadSymMOD(FileReader &file, ModLoadingFlags loadFlags)
 								if(patternSpeed > 1 && chnState.curPitchSlideAmt >= (patternSpeed - 1))
 								{
 									uint8 slideAmt = std::min<uint8>(0xDF, mpt::saturate_round<uint8>(chnState.curPitchSlideAmt / (patternSpeed - 1)));
-									chnState.curPitchSlideAmt -= slideAmt * (patternSpeed - 1);
+									chnState.curPitchSlideAmt -= static_cast<float>(slideAmt * (patternSpeed - 1));
 									// normal slide up
 									m.command = CMD_PORTAMENTOUP;
 									m.param = slideAmt;
 								} else if(chnState.curPitchSlideAmt >= 1.0f)
 								{
 									uint8 slideAmt = std::min<uint8>(15, mpt::saturate_round<uint8>(chnState.curPitchSlideAmt));
-									chnState.curPitchSlideAmt -= slideAmt;
+									chnState.curPitchSlideAmt -= static_cast<float>(slideAmt);
 									// fine slide up
 									m.command = CMD_PORTAMENTOUP;
 									m.param = slideAmt | 0xF0;
 								} else if(patternSpeed > 1 && chnState.curPitchSlideAmt <= -(patternSpeed - 1))
 								{
 									uint8 slideAmt = std::min<uint8>(0xDF, mpt::saturate_round<uint8>(-chnState.curPitchSlideAmt / (patternSpeed - 1)));
-									chnState.curPitchSlideAmt += slideAmt * (patternSpeed - 1);
+									chnState.curPitchSlideAmt += static_cast<float>(slideAmt * (patternSpeed - 1));
 									// normal slide down
 									m.command = CMD_PORTAMENTODOWN;
 									m.param = slideAmt;
 								} else if(chnState.curPitchSlideAmt <= -1.0f)
 								{
 									uint8 slideAmt = std::min<uint8>(14, mpt::saturate_round<uint8>(-chnState.curPitchSlideAmt));
-									chnState.curPitchSlideAmt += slideAmt;
+									chnState.curPitchSlideAmt += static_cast<float>(slideAmt);
 									// fine slide down
 									m.command = CMD_PORTAMENTODOWN;
 									m.param = slideAmt | 0xF0;
@@ -1849,13 +1849,13 @@ bool CSoundFile::ReadSymMOD(FileReader &file, ModLoadingFlags loadFlags)
 								if(patternSpeed > 1 && chnState.curPitchSlideAmt / 4 >= (patternSpeed - 1))
 								{
 									uint8 slideAmt = std::min<uint8>(9, mpt::saturate_round<uint8>(chnState.curPitchSlideAmt / (patternSpeed - 1)) / 4);
-									chnState.curPitchSlideAmt -= slideAmt * (patternSpeed - 1) * 4;
+									chnState.curPitchSlideAmt -= static_cast<float>(slideAmt * (patternSpeed - 1) * 4);
 									m.volcmd = VOLCMD_PORTAUP;
 									m.vol = slideAmt;
 								} else if(patternSpeed > 1 && chnState.curPitchSlideAmt / 4 <= -(patternSpeed - 1))
 								{
 									uint8 slideAmt = std::min<uint8>(9, mpt::saturate_round<uint8>(-chnState.curPitchSlideAmt / (patternSpeed - 1)) / 4);
-									chnState.curPitchSlideAmt += slideAmt * (patternSpeed - 1) * 4;
+									chnState.curPitchSlideAmt += static_cast<float>(slideAmt * (patternSpeed - 1) * 4);
 									m.volcmd = VOLCMD_PORTADOWN;
 									m.vol = slideAmt;
 								}
